requirements.txt
easydict==1.13
elasticsearch==7.12.0
face_recognition==1.3.0
ftfy==6.3.1
gdown==5.2.0
matplotlib==3.10.8
nltk==3.9.1
numpy==2.3.5
opencv_python==4.11.0.86
Pillow==12.0.0
plotly==6.5.0
PyYAML==6.0.2
PyYAML==6.0.3
regex==2024.11.6
Requests==2.32.5
scipy==1.16.3
streamlit==1.50.0
torch==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
ultralytics==8.3.232
yacs==0.1.8
